+++ LightDrive/workspace_file_manager.py
@@ -1,10 +1,11 @@
+from PySide6.QtWidgets import QFileDialog
 import tempfile
 import tarfile
 import shutil
 import json
 import os
 
-def read_workspace_file(workspace_file_path: str) -> tuple[dict, dict]:
+def read_workspace_file(workspace_file_path: str) -> tuple[dict, dict, dict]:
     """
     Reads a workspace file and returns its contents.
     :param workspace_file_path: The path to the workspace file
@@ -67,3 +68,115 @@ def write_workspace_file(workspace_file_path: str, fixtures: list, dmx_output_co
 
     if os.path.exists(tmp_dir):
         shutil.rmtree(tmp_dir)
+
+class WorkspaceFileManager:
+    def __init__(self, window, app, reboot_exit_code, current_workspace_file) -> None:
+        self.window = window
+        self.app = app
+        self.reboot_exit_code = reboot_exit_code
+        self.current_workspace_file = current_workspace_file
+
+    def new_workspace(self):
+        self.current_workspace_file = None
+        self.app.exit(self.reboot_exit_code)
+
+    def save_workspace_as(self):
+        dlg = QFileDialog(self.window, directory=os.path.expanduser("~"))
+        dlg.setNameFilter("Workspace (*.ldw)")
+        dlg.setDefaultSuffix(".ldw")
+        dlg.setAcceptMode(QFileDialog.AcceptSave)
+        if dlg.exec():
+            filename = dlg.selectedFiles()[0]
+            self.current_workspace_file = filename
+            self.save_workspace()
+
+    def save_workspace(self):
+        if not self.current_workspace_file:
+            self.save_workspace_as()
+        else:
+            snippet_configuration = self.get_snippet_configuration()
+            write_workspace_file(workspace_file_path=self.current_workspace_file,
+                                 fixtures=self.window.available_fixtures,
+                                 dmx_output_configuration=self.window.dmx_output.output_configuration,
+                                 snippet_configuration=snippet_configuration)
+
+    def get_snippet_configuration(self):
+        snippet_selector = self.window.ui.snippet_selector_tree
+        snippet_configuration = {}
+
+        def add_directory_content(item):
+            if item.extra_data["type"] == "directory":
+                item.extra_data["content"] = []
+                for j in range(item.childCount()):
+                    child = item.child(j)
+                    item.extra_data["content"].append(child.extra_data)
+                    add_directory_content(child)
+
+        for i in range(snippet_selector.topLevelItemCount()):
+            item = snippet_selector.topLevelItem(i)
+            snippet_configuration[str(i)] = item.extra_data
+            add_directory_content(item)
+        return snippet_configuration
+
+    def show_open_workspace_dialog(self):
+        dlg = QFileDialog(self.window, directory=os.path.expanduser("~"))
+        dlg.setNameFilter("Workspace (*.ldw)")
+        dlg.setDefaultSuffix(".ldw")
+        dlg.setFileMode(QFileDialog.ExistingFile)
+        if dlg.exec():
+            self.current_workspace_file = dlg.selectedFiles()[0]
+            self.app.exit(self.reboot_exit_code)  # Restart application (opens workspace while opening)
+
+    def open_workspace(self, workspace_file_path):
+        fixtures, dmx_output_configuration, snippets = read_workspace_file(workspace_file_path)
+        # Add the fixtures
+        for fixture in fixtures:
+            # Read the fixture data
+            fixture_dir = os.getenv('XDG_CONFIG_HOME', default=os.path.expanduser('~/.config')) + '/LightDrive/fixtures/'
+            with open(os.path.join(fixture_dir, fixture["id"] + ".json")) as f:
+                fixture_data = json.load(f)
+            # Add the fixture
+            self.window.add_fixture(amount = 1,
+                             fixture_data = fixture_data,
+                             universe = fixture["universe"],
+                             address = fixture["address"])
+
+        # Configure the dmx output
+        self.window.dmx_output.write_universe_configuration(dmx_output_configuration)
+
+        # Add the snippets
+        def add_snippets_to_parent(snippets, parent):
+            for snippet in snippets:
+                match snippet["type"]:
+                    case "cue":
+                        self.window.snippet_manager.snippet_create_cue(extra_data=snippet, parent=parent)
+                    case "scene":
+                        self.window.snippet_manager.snippet_create_scene(extra_data=snippet, parent=parent)
+                    case "efx_2d":
+                        self.window.snippet_manager.snippet_create_efx_2d(extra_data=snippet, parent=parent)
+                    case "rbg_matrix":
+                        self.window.snippet_manager.snippet_create_rgb_matrix(extra_data=snippet, parent=parent)
+                    case "script":
+                        self.window.snippet_manager.snippet_create_script(extra_data=snippet, parent=parent)
+                    case "directory":
+                        new_parent = self.window.snippet_manager.snippet_create_dir(extra_data=snippet, parent=parent)
+                        if "content" in snippet:
+                            add_snippets_to_parent(snippet["content"], new_parent)
+
+        for i, snippet in snippets.items():
+            print(i, snippet)
+            match snippet["type"]:
+                case "cue":
+                    self.window.snippet_manager.snippet_create_cue(extra_data=snippet)
+                case "scene":
+                    self.window.snippet_manager.snippet_create_scene(extra_data=snippet)
+                case "efx_2d":
+                    self.window.snippet_manager.snippet_create_efx_2d(extra_data=snippet)
+                case "rbg_matrix":
+                    self.window.snippet_manager.snippet_create_rgb_matrix(extra_data=snippet)
+                case "script":
+                    self.window.snippet_manager.snippet_create_script(extra_data=snippet)
+                case "directory":
+                    parent = self.window.snippet_manager.snippet_create_dir(extra_data=snippet)
+                    if "content" in snippet:
+                        add_snippets_to_parent(snippet["content"], parent)
